refactor : type 클래스들을 패키징 하기
