+++ src/main/java/com/example/rentalSystem/domain/email/controller/EmailController.java
@@ -6,7 +6,7 @@ import com.example.rentalSystem.domain.email.controller.dto.request.EmailRequest
 import com.example.rentalSystem.domain.email.controller.dto.response.EmailVerificationResult;
 import com.example.rentalSystem.domain.email.service.EmailService;
 import com.example.rentalSystem.global.response.ApiResponse;
-import com.example.rentalSystem.global.response.SuccessType;
+import com.example.rentalSystem.global.response.type.SuccessType;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -15,9 +15,9 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
-@RequestMapping("/email")
 @RequiredArgsConstructor
-public class EmailController {
+@RequestMapping("/email")
+public class EmailController implements EmailControllerDocs {
 
     private final EmailService emailService;
 
@@ -34,7 +34,8 @@ public class EmailController {
     }
 
     @PostMapping("/check-code")
-    public ApiResponse<?> checkCode(@RequestBody AuthCodeRequest authCodeRequest) {
+    public ApiResponse<EmailVerificationResult> checkCode(
+        @RequestBody AuthCodeRequest authCodeRequest) {
         EmailVerificationResult emailVerificationResult = emailService.verificationCode(
             authCodeRequest.email(), authCodeRequest.authCode());
         return ApiResponse.success(SuccessType.SUCCESS, emailVerificationResult);
