+++ src/main/java/com/example/rentalSystem/domain/email/implement/MailLoader.java
@@ -2,7 +2,7 @@ package com.example.rentalSystem.domain.email.implement;
 
 import com.example.rentalSystem.domain.email.repository.MailRepository;
 import com.example.rentalSystem.global.exception.custom.CustomException;
-import com.example.rentalSystem.global.response.ErrorType;
+import com.example.rentalSystem.global.response.type.ErrorType;
 import java.util.Objects;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
