+++ src/main/java/com/example/rentalSystem/domain/member/entity/CustomerDetails.java
@@ -2,7 +2,7 @@ package com.example.rentalSystem.domain.member.entity;
 
 import com.example.rentalSystem.domain.student.entity.Student;
 import com.example.rentalSystem.global.exception.custom.CustomException;
-import com.example.rentalSystem.global.response.ErrorType;
+import com.example.rentalSystem.global.response.type.ErrorType;
 import java.util.Arrays;
 import java.util.Collection;
 import java.util.stream.Collectors;
