+++ src/main/java/com/example/rentalSystem/domain/professor/controller/ProfessorController.java
@@ -3,8 +3,7 @@ package com.example.rentalSystem.domain.professor.controller;
 import com.example.rentalSystem.domain.professor.controller.dto.response.ProfessorDetailResponse;
 import com.example.rentalSystem.domain.professor.service.ProfessorService;
 import com.example.rentalSystem.global.response.ApiResponse;
-import com.example.rentalSystem.global.response.SuccessType;
-import java.util.List;
+import com.example.rentalSystem.global.response.type.SuccessType;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
