+++ src/main/java/com/example/rentalSystem/domain/rentalhistory/implement/ProfessorHistoryImpl.java
@@ -4,7 +4,7 @@ import com.example.rentalSystem.domain.rentalhistory.entity.ProfessorHistory;
 import com.example.rentalSystem.domain.rentalhistory.entity.RentalHistory;
 import com.example.rentalSystem.domain.rentalhistory.repository.ProfessorHistoryRepository;
 import com.example.rentalSystem.global.exception.custom.CustomException;
-import com.example.rentalSystem.global.response.ErrorType;
+import com.example.rentalSystem.global.response.type.ErrorType;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
