--- src/main/java/com/example/rentalSystem/global/response/ResultType.java
@@ -1,6 +1,0 @@
-package com.example.rentalSystem.global.response;
-
-public enum ResultType {
-
-  SUCCESS, FAIL
-}
