+++ src/main/java/com/example/rentalSystem/global/response/type/ResultType.java
@@ -1,0 +1,6 @@
+package com.example.rentalSystem.global.response.type;
+
+public enum ResultType {
+
+    SUCCESS, FAIL
+}
